+++ modgeo/control/Conexao.php
@@ -1,3 +1,4 @@
+<?php
 class Conexao
 {
     static function conectar()
