--- view/CadNomeSondagem.php
@@ -1,12 +1,0 @@
-<form action="../control/NomeSondagemControl.php" method="POST">
-    Nome<input type="text" name = "nome"><br>
-    Coordenada X<input type="text" name = "coordx"><br>
-    Coordenada Y<input type="text" name = "coordy"><br>
-    Profundidade<input type="text" name = "profundidade"><br>
-    Inclinação<input type="text" name = "inclinacao"><br>
-    Georeferenciamento<input type="text" name = "georeferenciamento"><br>
-    Cota<input type="text" name = "cota"><br>
-    <input type="hidden" name="acao" value="1">
-    <input type="submit" value="gravar">
-
-</form>
